refactor(json): move json-coder and adapter-contract into 'json' package
