+++ infrastructure/src/main/java/io/github/aivruu/teams/shared/infrastructure/json/JsonCodecAdapterContract.java
@@ -14,7 +14,7 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
-package io.github.aivruu.teams.shared.infrastructure.adapter;
+package io.github.aivruu.teams.shared.infrastructure.json;
 
 import com.google.gson.JsonDeserializer;
 import com.google.gson.JsonSerializer;
